refactor: unused styles
