--- frontend/src/styles/pages/boards/newRegularBoard.styles.tsx
@@ -1,11 +1,0 @@
-import { styled } from '@/styles/stitches/stitches.config';
-
-const ContentSelectContainer = styled('section', {
-  display: 'flex',
-  width: '100%',
-  minHeight: 'calc(100vh - $141)',
-  alignItems: 'center',
-  justifyContent: 'center',
-});
-
-export { ContentSelectContainer };
