+++ frontend/src/styles/pages/pages.styles.tsx
@@ -1,0 +1,7 @@
+import Flex from '@/components/Primitives/Layout/Flex/Flex';
+import { styled } from '../stitches/stitches.config';
+
+export const StyledForm = styled('form', Flex, {
+  flex: '1 1 auto',
+  padding: '64px 92px 57px 152px',
+});
